requirements.txt
Django>=5.2.1
psycopg2-binary
celery
redis
djangorestframework
django-celery-results
